requirements.txt
chromadb
sentence-transformers
PyMuPDF
python-docx
openai
fastapi
uvicorn
